allow formatting for slack or CLI
